+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Implement MAP()
 * Implement REDUCE()
 * Implement ORV()
+* Implement TEXTJOIN()
 
 
 ## xel.js 1.1.1  released 2024-03-30  (Nami)